Rename project and package from "My Calisthenics" to "Apollo"
* Rename package from `com.example.mycalisthenics` to `com.alois.apollo` across all source files and tests
* Update `applicationId` and `namespace` in `build.gradle.kts`
* Update application name in `strings.xml` and UI components
* Rename `MyCalisthenicsTheme` to `ApolloTheme`
* Remove unused comments and redundant code in `MainScreen.kt`, `Theme.kt`, and `WorkoutViewModel.kt`
* Refactor imports to reflect new package structure
