+++ app/src/androidTest/java/com/alois/apollo/ExampleInstrumentedTest.kt
@@ -1,12 +1,10 @@
-package com.example.mycalisthenics
+package com.alois.apollo
 
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
-
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
@@ -19,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.mycalisthenics", appContext.packageName)
+        assertEquals("com.example.com.alois.apollo", appContext.packageName)
     }
 }
