+++ app/src/main/java/com/alois/apollo/MainActivity.kt
@@ -1,20 +1,28 @@
-package com.example.mycalisthenics
+package com.alois.apollo
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.runtime.*
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.mycalisthenics.ui.*
-import com.example.mycalisthenics.ui.theme.MyCalisthenicsTheme
+import com.alois.apollo.ui.MainScreen
+import com.alois.apollo.ui.ReadOnlyRecapScreen
+import com.alois.apollo.ui.RecapScreen
+import com.alois.apollo.ui.WorkoutScreen
+import com.alois.apollo.ui.WorkoutViewModel
+import com.alois.apollo.ui.theme.ApolloTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
-            MyCalisthenicsTheme {
+            ApolloTheme {
                 val viewModel: WorkoutViewModel = viewModel()
                 val activeWorkout by viewModel.activeWorkout.collectAsState()
                 val showRecap by viewModel.showRecap.collectAsState()
