+++ app/src/main/java/com/alois/apollo/data/local/WorkoutDao.kt
@@ -1,9 +1,9 @@
-package com.example.mycalisthenics.data.local
+package com.alois.apollo.data.local
 
 import androidx.room.Dao
+import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
-import androidx.room.Delete
 
 @Dao
 interface WorkoutDao {
