+++ app/src/main/java/com/alois/apollo/ui/MainScreen.kt
@@ -1,24 +1,49 @@
-package com.example.mycalisthenics.ui
+package com.alois.apollo.ui
 
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.History
 import androidx.compose.material.icons.filled.PlayArrow
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.ListItem
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SwipeToDismissBox
+import androidx.compose.material3.SwipeToDismissBoxValue
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.rememberSwipeToDismissBoxState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.mycalisthenics.data.local.WorkoutSession
+import com.alois.apollo.data.local.WorkoutSession
 import java.text.SimpleDateFormat
-import java.util.*
+import java.util.Date
+import java.util.Locale
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -29,11 +54,11 @@ fun MainScreen(
 ) {
     val workouts by viewModel.availableWorkouts.collectAsState()
     val history by viewModel.history.collectAsState()
-    val isFrench = viewModel.isFrench()
+    viewModel.isFrench()
 
     Scaffold(
         topBar = {
-            TopAppBar(title = { Text("My Calisthenics") })
+            TopAppBar(title = { Text("Apollo") })
         }
     ) { paddingValues ->
         LazyColumn(
@@ -106,15 +131,11 @@ fun SwipeToDeleteSession(
 ) {
     val dismissState = rememberSwipeToDismissBoxState()
     
-    if (dismissState.currentValue == SwipeToDismissBoxValue.EndToStart) {
-        // Confirmation is implicit or handled by button revealing
-    }
-
     SwipeToDismissBox(
         state = dismissState,
         enableDismissFromStartToEnd = false,
         backgroundContent = {
-            val color = if (dismissState.dismissDirection == SwipeToDismissBoxValue.EndToStart) {
+            if (dismissState.dismissDirection == SwipeToDismissBoxValue.EndToStart) {
                 MaterialTheme.colorScheme.errorContainer
             } else {
                 MaterialTheme.colorScheme.surface
